desktop_app/requirements.txt
customtkinter
ffmpeg-python
pyinstaller
pillow
tkinterdnd2
